+++ Gruntfile.js
@@ -67,12 +67,17 @@ module.exports = function(grunt) {
           files: files
         };
       }())
+    },
+    watch: {
+      files: ['src/**/*.js'],
+      tasks: ['default']
     }
   });
 
   grunt.loadNpmTasks('grunt-contrib-jshint');
   grunt.loadNpmTasks('grunt-contrib-uglify');
   grunt.loadNpmTasks('grunt-contrib-concat');
+  grunt.loadNpmTasks('grunt-contrib-watch');
 
   // Default task.
   grunt.registerTask('default', ['jshint', 'concat', 'uglify']);