+++ src/validatorerrors.js
@@ -3,6 +3,11 @@ var ValidatorErrors = function() {};
 ValidatorErrors.prototype = {
 	constructor: ValidatorErrors,
 
+	/**
+	 * returns an array of error messages for an attribute, or an empty array
+	 * @param  {String} attribute A key in the data object being validated
+	 * @return {Array}           	An array of error messages
+	 */
 	get: function(attribute) {
 		if (this[attribute]) {
 			return this[attribute];	
@@ -11,11 +16,24 @@ ValidatorErrors.prototype = {
 		return [];
 	},
 
+	/**
+	 * returns the first error message for an attribute, false otherwise
+	 * @param  {String} attribute A key in the data object being validated
+	 * @return {String}           First error message or false
+	 */
 	first: function(attribute) {
 		if (this[attribute]) {
 			return this[attribute][0];	
 		}
 		
 		return false;
+	},
+
+	/**
+	 * Get all error messages from all failing attributes
+	 * @return {Object} Failed attribute names for keys and an array of messages for values
+	 */
+	all: function() {
+		return this;
 	}
 };