+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Adds Validator.make() as an alias to new Validator()
 * Broke up library into smaller files that get combined during build
+* Replaces jasmine-node with Karma and Grunt watch for testing
+* Validator errors object has an all() method to get all error messages
+* Adds Istanbul for test coverage reporting
 
 ### 1.1.0